+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 2.101.0
 * Add more Venmo error codes to `ValidationErrorCode`
+* Add `threeDSecureInfo` to `CreditCardVerification`
+* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 
 ## 2.100.0
 * Add `CustomActionsPaymentMethod` support
@@ -19,7 +21,6 @@
 * Add `transactionFeeCurrencyIsoCode` field to `LocalPaymentDetail`
 * Add `refundFromTransactionFeeAmount` field to `LocalPaymentDetail`
 * Add `refundFromTransactionFeeCurrencyIsoCode` field to `LocalPaymentDetail`
-* Add `xid`, `cavv`, `eciFlag`, `dsTransactionId`, and `threeDSecureVersion` to `ThreeDSecureInfo`
 * Add `dsTransactionId` to `ThreeDSecurePassthruRequest`
 * Add `PayPalHereDetails` to `Transaction`